rippleを共通化
